Initial conv2d support (#174)
* minor changes to support conv2d

* add flatten_either utility function

* initial conv2d support

* update tests

* fix for wasm, gcc10, clang13

* fix bias

* fix test for gcc & clang CI

* fix for clang werror T_T

+++ include/nmtools/array/index/count.hpp
@@ -27,7 +27,7 @@ namespace nmtools::index
         }
         else {
             auto n = len(array);
-            for (size_t i=0; i<n; i++)
+            for (size_t i=0; i<(size_t)n; i++)
                 count_impl(i);
         }
